requirements.txt
selenium
json
google-cloud-translate==2.0.1